CMake: Adding BUILD_WITH_THREADS  option
Compared to the 4.0 branch, there is the instruction to link against pthread now.
However, pthreads is (as it is named) only available on posix (except you are using MinGW on Windows).
Anyway, I disabled hereby the usage of (p)threads on Windows. Probably there is a more elegant way to handle this situation, but that will allow Windows builds again at least.