Detect pthread quietly and adding it as found
I'm doing basically the same as done in FindGTest. FindGTest is developed by CMake and there they are looking for Threads quietly.
That happens there with line 200 and following. Then they are checking whether it has been found and if that is the case, set_target_properties is being used. I'm left-handed at this point, so I implemented something that looks similar to what I (think to) see, but I trust more.