Fix crash occuring when the faces have wrong vertex ids
It can happen that the triangles in the object-to-be-loaded refer to vertex id's that do not exist
in the vertices table. This means that either Cura made a mistake in saving these objects or that
the triangles of the objects have been manually changed. In any case, this was causing a crash.

This is now fixed by valdiating whether the vertex id's are within limits. Even if one of them is
not, it means the triangle will not be loaded and we will end up with a possibly non-manifold and
overall wrong object (especially if there are multiple mistakes like that). So, in such cases,
refrain from loading the object by returning an empty bytearray.

CURA-7790