Don't filter metadata entries with unknown namespace
We'll just remove the cura: namespace then, to make Cura's settings still parse correctly.
This is a compromise. Basically this removes any 'cura:' namespace even though it is not a real namespace. This is not conformant to 3MF which would say that these names need to be preserved. This means that we'd need to pass on the 'cura:' namespace to Cura itself and let them handle it. However the Cura front-end code doesn't know the namespace aliases any more. It doesn't have access to them from there, since they are not actual XML namespaces but just a convention of the name attribute of metadata tags we have. The front-end doesn't even have access to the XML, let alone the namespaces in there for each tag.
The front-end needs to add the 'cura' namespace back if it recognises the name as a setting key, when writing the 3MF.

Contributes to issue CURA-7615.