Use Conan for dependencies
This change allows you to use Conan as dependency
manager for SIP, pugimxl and gtest.
**Completely optional!!*

It will either download a pre-build binary if available for the architecture
and used options, or build it (use --build=missing in Clion conan settings)
on time and store it in the conan cache. It will then create the appropriate
cmake find scripts.

**NOTE:**
Use the Conan script for SIP https://github.com/jellespijker/conan-sip
to create the SIP dependency for now. This should first be generated
locally since it hasn't been added to a remote repo yet.